Implement tab design from antd library, and move some contents into newly created Overview.js file
